+++ src/Console/Commands/GenerateBlueprint.php
@@ -235,7 +235,7 @@ class GenerateBlueprint extends Command
 
         Blueprint::make($resource->handle())
             ->setContents([
-                'sections' => [
+                'tabs' => [
                     'main' => ['fields' => $mainSection],
                     'sidebar' => ['fields' => $sidebarSection],
                 ],