+++ tests/TestCase.php
@@ -106,7 +106,7 @@ abstract class TestCase extends OrchestraTestCase
                 Post::class => [
                     'name' => 'Posts',
                     'blueprint' => [
-                        'sections' => [
+                        'tabs' => [
                             'main' => [
                                 'fields' => [
                                     [
@@ -169,7 +169,7 @@ abstract class TestCase extends OrchestraTestCase
                 Author::class => [
                     'name' => 'Author',
                     'blueprint' => [
-                        'sections' => [
+                        'tabs' => [
                             'main' => [
                                 'fields' => [
                                     [